analytics_api/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
numpy
pandas
